ui: 화면세로스크롤 적용
