+++ makeat_fe/pubspec.yaml
@@ -43,7 +43,9 @@ dependencies:
   flutter_secure_storage: ^8.0.0
   http: ^0.13.5
   kakao_flutter_sdk: ^1.4.2
-  flutter_webview_plugin: ^0.4.0
+  flutter_webview_plugin: 
+    git:
+      url: https://github.com/hardy716/flutter_webview_plugin
 #  flutter_webview_plugin:
 #    path: ../../flutter_webview_plugin
 webview_flutter: ^4.0.7
